use percentCorrect instead of summaryStat, extract out large chunks of redux.js
